refactor: reorganize core PR functions into modular structure
