--- src/core/pr.ts
@@ -1,116 +1,0 @@
-import { DateTime } from 'luxon';
-
-import { fetchAllPRListInPeriod } from '../api';
-import { PRQueryParams, PR } from '../types';
-
-export function uniqueArray<T>(array: T[]): T[] {
-  return Array.from(new Set(array));
-}
-
-export async function getUserPRListByCreationAndParticipation(
-  options: PRQueryParams,
-): Promise<{ userCreatedPRList: PR[]; userParticipatedPRList: PR[] }> {
-  let allPRList = await fetchAllPRListInPeriod(options);
-  if (options.targetBranch) {
-    allPRList = filterPRListByTargetBranch(allPRList, options.targetBranch);
-  }
-
-  const createdPRList: PR[] = [];
-  const participatedPRList: PR[] = [];
-
-  allPRList.forEach((pr) => {
-    if (pr.author === options.username) {
-      createdPRList.push({
-        ...pr,
-        comments: uniqueArray(pr.comments ?? []),
-        reviewers: uniqueArray(pr.reviewers),
-        createdAt: convertUTCISOToLocal(pr.createdAt),
-        mergedAt: pr.mergedAt ? convertUTCISOToLocal(pr.mergedAt) : 'Not merged yet',
-      });
-    } else if (pr.comments?.includes(options.username) || pr.reviewers.includes(options.username)) {
-      participatedPRList.push({
-        ...pr,
-        comments: uniqueArray(pr.comments ?? []),
-        reviewers: uniqueArray(pr.reviewers),
-        createdAt: convertUTCISOToLocal(pr.createdAt),
-        mergedAt: pr.mergedAt ? convertUTCISOToLocal(pr.mergedAt) : 'Not merged yet',
-      });
-    }
-  });
-
-  return { userCreatedPRList: createdPRList, userParticipatedPRList: participatedPRList };
-}
-
-export async function getUserCreatedPRListInPeriod(options: PRQueryParams): Promise<PR[]> {
-  let allPRList = await fetchAllPRListInPeriod(options);
-  if (options.targetBranch) {
-    allPRList = filterPRListByTargetBranch(allPRList, options.targetBranch);
-  }
-
-  return allPRList
-    .filter((pr) => pr.author === options.username)
-    .map((pr) => ({
-      ...pr,
-      comments: uniqueArray(pr.comments ?? []),
-      reviewers: uniqueArray(pr.reviewers),
-      createdAt: convertUTCISOToLocal(pr.createdAt),
-      mergedAt: pr.mergedAt ? convertUTCISOToLocal(pr.mergedAt) : 'Not merged yet',
-    }));
-}
-
-export async function getUserParticipatedPRListInPeriod(options: PRQueryParams): Promise<PR[]> {
-  let allPRList = await fetchAllPRListInPeriod(options);
-  if (options.targetBranch) {
-    allPRList = filterPRListByTargetBranch(allPRList, options.targetBranch);
-  }
-
-  const participatedPRList = allPRList
-    .filter(
-      (pr) => pr.comments?.includes(options.username) || pr.reviewers.includes(options.username),
-    )
-    .map((pr) => ({
-      ...pr,
-      comments: uniqueArray(pr.comments ?? []),
-      reviewers: uniqueArray(pr.reviewers),
-      createdAt: convertUTCISOToLocal(pr.createdAt),
-      mergedAt: pr.mergedAt ? convertUTCISOToLocal(pr.mergedAt) : 'Not merged yet',
-    }));
-
-  return participatedPRList;
-}
-
-export function getUserCreatedPRCountInPeriod(prList: PR[]): number {
-  return prList.length;
-}
-
-export async function getUserCreatedPRListInPeriodByLabel(
-  prList: PR[],
-  labelFilter: string[],
-): Promise<PR[]> {
-  return prList.filter((pr) =>
-    labelFilter.some((filter) =>
-      pr.labels.some((label) => label.toLowerCase().includes(filter.toLowerCase())),
-    ),
-  );
-}
-
-export function getUserPRCountByLabelInPeriod(prList: PR[]): Record<string, number> {
-  const labelCountMap: Record<string, number> = {};
-  prList.map((pr) => {
-    pr.labels.forEach((label) => {
-      labelCountMap[label] = (labelCountMap[label] || 0) + 1;
-    });
-  });
-  return labelCountMap;
-}
-
-export function filterPRListByTargetBranch(prList: PR[], targetBranch: string): PR[] {
-  const parsedTargetBranch = targetBranch.includes(':') ? targetBranch.split(':')[1] : targetBranch;
-  return prList.filter((pr) => pr.targetBranch === parsedTargetBranch);
-}
-
-export function convertUTCISOToLocal(date: string): string {
-  const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-  const local = DateTime.fromISO(date, { zone: timeZone });
-  return local.toFormat('yyyy-MM-dd HH:mm:ss');
-}
