+++ src/core/pr/filters.ts
@@ -1,0 +1,6 @@
+import { PR } from '../../types';
+
+export function filterPRListByTargetBranch(prList: PR[], targetBranch: string): PR[] {
+  const parsedTargetBranch = targetBranch.includes(':') ? targetBranch.split(':')[1] : targetBranch;
+  return prList.filter((pr) => pr.targetBranch === parsedTargetBranch);
+}
